Replace `otel4s-testkit` with `otel4s-oteljava-testkit`
