+++ .github/workflows/ci.yml
@@ -88,11 +88,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p testkit/metrics/jvm/target oteljava/metrics/target testkit/common/jvm/target core/trace/.js/target semconv/.jvm/target core/common/.jvm/target unidocs/target core/metrics/.native/target core/all/.native/target core/metrics/.jvm/target core/all/.js/target core/metrics/.js/target core/all/.jvm/target core/trace/.native/target semconv/.js/target oteljava/common/target scalafix/rules/target core/trace/.jvm/target core/common/.native/target core/common/.js/target oteljava/trace/target semconv/.native/target testkit/all/jvm/target oteljava/all/target project/target
+        run: mkdir -p oteljava/metrics/target core/trace/.js/target semconv/.jvm/target core/common/.jvm/target oteljava/common-testkit/target unidocs/target core/metrics/.native/target core/all/.native/target core/metrics/.jvm/target core/all/.js/target core/metrics/.js/target core/all/.jvm/target core/trace/.native/target oteljava/metrics-testkit/target semconv/.js/target oteljava/common/target scalafix/rules/target core/trace/.jvm/target core/common/.native/target core/common/.js/target oteljava/trace/target oteljava/testkit/target semconv/.native/target oteljava/all/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar testkit/metrics/jvm/target oteljava/metrics/target testkit/common/jvm/target core/trace/.js/target semconv/.jvm/target core/common/.jvm/target unidocs/target core/metrics/.native/target core/all/.native/target core/metrics/.jvm/target core/all/.js/target core/metrics/.js/target core/all/.jvm/target core/trace/.native/target semconv/.js/target oteljava/common/target scalafix/rules/target core/trace/.jvm/target core/common/.native/target core/common/.js/target oteljava/trace/target semconv/.native/target testkit/all/jvm/target oteljava/all/target project/target
+        run: tar cf targets.tar oteljava/metrics/target core/trace/.js/target semconv/.jvm/target core/common/.jvm/target oteljava/common-testkit/target unidocs/target core/metrics/.native/target core/all/.native/target core/metrics/.jvm/target core/all/.js/target core/metrics/.js/target core/all/.jvm/target core/trace/.native/target oteljava/metrics-testkit/target semconv/.js/target oteljava/common/target scalafix/rules/target core/trace/.jvm/target core/common/.native/target core/common/.js/target oteljava/trace/target oteljava/testkit/target semconv/.native/target oteljava/all/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
