--- testkit/common/shared/src/main/scala/org/typelevel/otel4s/testkit/Implicits.scala
@@ -1,31 +1,0 @@
-/*
- * Copyright 2022 Typelevel
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.typelevel.otel4s
-package testkit
-
-import cats.Hash
-import cats.Show
-
-object Implicits {
-
-  implicit val attributeListHash: Hash[List[Attribute[_]]] =
-    Hash.fromUniversalHashCode
-
-  implicit val attributeListShow: Show[List[Attribute[_]]] =
-    Show.fromToString
-
-}
